+++ src/main/java/org/example/entities/ArticuloInsumo.java
@@ -1,0 +1,16 @@
+package org.example.entities;
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@ToString
+public class ArticuloInsumo extends Articulo {
+    private double precioCompra;
+    private int stockActual;
+    private int stockMaximo;
+    private boolean esParaElaborar;
+
+}
