+++ src/main/java/org/example/entities/ArticuloManufacturadoDetalle.java
@@ -1,0 +1,16 @@
+package org.example.entities;
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@ToString
+public class ArticuloManufacturadoDetalle {
+    private int cantidad;
+
+    private ArticuloInsumo articuloInsumo;
+
+
+}
