+++ src/main/java/org/example/entities/enums/Estado.java
@@ -1,3 +1,5 @@
+package org.example.entities.enums;
+
 public enum Estado {
     PREPARACION,
     PENDIENTE,
